// packages/openapi-bench/generators/openapi-generator/client/model/weird-names.ts
/**
 * OpenAPI Bench API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WeirdNames { 
    'x-request-id': string;
    '@type': string;
    $ref?: string;
    '1stPlace'?: boolean;
    snake_case_prop?: string;
    'with space'?: string;
    'class'?: string;
    'default'?: string;
    'delete'?: boolean;
    constructor?: string;
    __proto_like?: string;
}